+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/RSMineshaftEndStructure.java
@@ -13,6 +13,7 @@ import net.minecraft.world.biome.provider.BiomeProvider;
 import net.minecraft.world.gen.ChunkGenerator;
 import net.minecraft.world.gen.Heightmap;
 import net.minecraft.world.gen.feature.NoFeatureConfig;
+import net.minecraft.world.gen.settings.StructureSeparationSettings;
 
 import java.util.List;
 import java.util.Objects;
@@ -35,11 +36,14 @@ public class RSMineshaftEndStructure extends RSMineshaftStructure {
 
     @Override
     protected boolean shouldStartAt(ChunkGenerator chunkGenerator, BiomeProvider biomeSource, long seed, SharedSeedRandom chunkRandom, int chunkX, int chunkZ, Biome biome, ChunkPos chunkPos, NoFeatureConfig featureConfig) {
-        chunkRandom.setLargeFeatureSeed(seed + Objects.requireNonNull(chunkGenerator.getStructuresConfig().getForType(this)).getSalt(), chunkX, chunkZ);
-        double d = (probability / 10000D);
-        if(chunkRandom.nextDouble() < d) {
-            int landHeight = chunkGenerator.func_222531_c(chunkX << 4, chunkZ << 4, Heightmap.Type.WORLD_SURFACE_WG);
-            return RepurposedStructures.RSMineshaftsConfig.barrensIslandsEndMineshafts.get() || landHeight > 20;
+        StructureSeparationSettings structureConfig = chunkGenerator.getStructuresConfig().getForType(this);
+        if(structureConfig != null){
+            chunkRandom.setLargeFeatureSeed(seed + structureConfig.getSalt(), chunkX, chunkZ);
+            double d = (probability / 10000D);
+            if(chunkRandom.nextDouble() < d) {
+                int landHeight = chunkGenerator.func_222531_c(chunkX << 4, chunkZ << 4, Heightmap.Type.WORLD_SURFACE_WG);
+                return RepurposedStructures.RSMineshaftsConfig.barrensIslandsEndMineshafts.get() || landHeight > 20;
+            }
         }
         return false;
     }