+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/RSMineshaftStructure.java
@@ -14,6 +14,7 @@ import net.minecraft.world.gen.feature.structure.Structure;
 import net.minecraft.world.gen.feature.structure.StructurePiece;
 import net.minecraft.world.gen.feature.structure.StructureStart;
 import net.minecraft.world.gen.feature.template.TemplateManager;
+import net.minecraft.world.gen.settings.StructureSeparationSettings;
 
 import java.util.Objects;
 
@@ -34,9 +35,13 @@ public class RSMineshaftStructure extends AbstractBaseStructure {
 
     @Override
     protected boolean shouldStartAt(ChunkGenerator chunkGenerator, BiomeProvider biomeSource, long seed, SharedSeedRandom chunkRandom, int x, int z, Biome biome, ChunkPos chunkPos, NoFeatureConfig featureConfig) {
-        chunkRandom.setLargeFeatureSeed(seed + Objects.requireNonNull(chunkGenerator.getStructuresConfig().getForType(this)).getSalt(), x, z);
-        double d = (probability / 10000D);
-        return chunkRandom.nextDouble() < d;
+        StructureSeparationSettings structureConfig = chunkGenerator.getStructuresConfig().getForType(this);
+        if(structureConfig != null) {
+            chunkRandom.setLargeFeatureSeed(seed + structureConfig.getSalt(), x, z);
+            double d = (probability / 10000D);
+            return chunkRandom.nextDouble() < d;
+        }
+        return false;
     }
 
     @Override